# backend/docs/swagger.json
{
  "basePath": "/api/v1",
  "definitions": {
    "controllers.JournalSummary": {
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "controllers.ReverseJournalRequest": {
      "properties": {
        "description": {
          "type": "string"
        }
      },
      "required": [
        "description"
      ],
      "type": "object"
    },
...
    "models.SSOTAccountBalance": {
      "properties": {
        "account_category": {
          "type": "string"
        },
        "account_code": {
          "type": "string"
        },
        "account_id": {
          "type": "integer"
        },
        "account_name": {
          "type": "string"
        },
        "account_type": {
          "type": "string"
        },
        "current_balance": {
          "type": "number"
        },
        "is_active": {
          "type": "boolean"
        },
        "is_header": {
          "type": "boolean"
        },
        "last_transaction_date": {
          "type": "string"
        },
        "last_updated": {
          "type": "string"
        },
        "normal_balance": {
          "type": "string"
        },
        "total_credits": {
          "type": "number"
        },
        "total_debits": {
          "type": "number"
        },
        "transaction_count": {
          "type": "integer"
        }
      },
      "type": "object"
    },
...
    "services.JournalEntryRequest": {
      "properties": {
        "auto_post": {
          "type": "boolean"
        },
        "created_by": {
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "entry_date": {
          "type": "string"
        },
        "lines": {
          "items": {
            "$ref": "#/definitions/services.JournalLineRequest"
          },
          "type": "array"
        },
        "reference": {
          "type": "string"
        },
        "source_id": {
          "type": "integer"
        },
        "source_type": {
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "services.JournalLineRequest": {
      "properties": {
        "account_id": {
          "type": "integer"
        },
        "credit_amount": {
          "type": "number"
        },
        "debit_amount": {
          "type": "number"
        },
        "description": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "services.JournalLineResponse": {
      "properties": {
        "account_id": {
          "type": "integer"
        },
        "credit_amount": {
          "type": "number"
        },
        "debit_amount": {
          "type": "number"
        },
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "line_number": {
          "type": "integer"
        }
      },
      "type": "object"
    },
    "services.JournalResponse": {
      "properties": {
        "created_at": {
          "type": "string"
        },
        "entry_number": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "is_balanced": {
          "type": "boolean"
        },
        "lines": {
          "items": {
            "$ref": "#/definitions/services.JournalLineResponse"
          },
          "type": "array"
        },
        "status": {
          "type": "string"
        },
        "total_credit": {
          "type": "number"
        },
        "total_debit": {
          "type": "number"
        },
        "updated_at": {
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "services.PaginatedJournalResponse": {
      "properties": {
        "data": {
          "items": {
            "$ref": "#/definitions/services.JournalResponse"
          },
          "type": "array"
        },
        "limit": {
          "type": "integer"
        },
        "page": {
          "type": "integer"
        },
        "total": {
          "type": "integer"
        },
        "total_pages": {
          "type": "integer"
        }
      },
      "type": "object"
    },
...
  },
  "host": "localhost:8080",
  "info": {
    "contact": {
      "email": "support@swagger.io",
      "name": "API Support",
      "url": "http://www.swagger.io/support"
    },
    "description": "API untuk aplikasi sistem akuntansi yang komprehensif dengan fitur lengkap manajemen keuangan, inventory, sales, purchases, dan reporting.",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService": "http://swagger.io/terms/",
    "title": "Sistema Akuntansi API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/journals": {
      "get": {
        "description": "Get journal entries with filtering, pagination, and sorting",
        "parameters": [
          {
            "description": "Source type filter",
            "in": "query",
            "name": "source_type",
            "type": "string"
          },
          {
            "description": "Status filter",
            "in": "query",
            "name": "status",
            "type": "string"
          },
          {
            "description": "Date from filter (YYYY-MM-DD)",
            "in": "query",
            "name": "date_from",
            "type": "string"
          },
          {
            "description": "Date to filter (YYYY-MM-DD)",
            "in": "query",
            "name": "date_to",
            "type": "string"
          },
          {
            "description": "Reference search",
            "in": "query",
            "name": "reference",
            "type": "string"
          },
          {
            "default": 1,
            "description": "Page number",
            "in": "query",
            "name": "page",
            "type": "integer"
          },
          {
            "default": 20,
            "description": "Items per page",
            "in": "query",
            "name": "limit",
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/services.PaginatedJournalResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Get journal entries list",
        "tags": [
          "Journal"
        ]
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Create a new journal entry with validation and optional auto-posting",
        "parameters": [
          {
            "description": "Journal Entry Request",
            "in": "body",
            "name": "journal",
            "required": true,
            "schema": {
              "$ref": "#/definitions/services.JournalEntryRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/services.JournalResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Create a new journal entry",
        "tags": [
          "Journal"
        ]
      }
    },
    "/api/v1/journals/account-balances": {
      "get": {
        "description": "Get account balances from materialized view",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/models.SSOTAccountBalance"
              },
              "type": "array"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Get account balances",
        "tags": [
          "Journal"
        ]
      }
    },
    "/api/v1/journals/account-balances/refresh": {
      "post": {
        "description": "Manually refresh the materialized view for account balances",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Refresh account balances",
        "tags": [
          "Journal"
        ]
      }
    },
    "/api/v1/journals/summary": {
      "get": {
        "description": "Get summary statistics for journal entries",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/controllers.JournalSummary"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Get journal summary",
        "tags": [
          "Journal"
        ]
      }
    },
    "/api/v1/journals/{id}": {
      "get": {
        "description": "Get a specific journal entry with its lines",
        "parameters": [
          {
            "description": "Journal Entry ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/services.JournalResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Get journal entry by ID",
        "tags": [
          "Journal"
        ]
      }
    },
    "/api/v1/journals/{id}/post": {
      "put": {
        "description": "Change journal entry status from draft to posted",
        "parameters": [
          {
            "description": "Journal Entry ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Post a journal entry",
        "tags": [
          "Journal"
        ]
      }
    },
    "/api/v1/journals/{id}/reverse": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Create a reversing journal entry for a posted entry",
        "parameters": [
          {
            "description": "Journal Entry ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "description": "Reversal Request",
            "in": "body",
            "name": "reversal",
            "required": true,
            "schema": {
              "$ref": "#/definitions/controllers.ReverseJournalRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/services.JournalResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Reverse a journal entry",
        "tags": [
          "Journal"
        ]
      }
    },
...
  },
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Type \"Bearer\" followed by a space and JWT token.",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}